requirements.txt
alabaster==0.7.13
asgiref==3.7.2
attrs==23.1.0
Babel==2.14.0
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.4
django-cors-headers==4.2.0
django-cuser==2017.3.16
django-environ==0.10.0
django-filter==23.2
django-jazzmin==2.6.0
django-middleware-global-request==0.3.1
django-modeltranslation==0.18.11
django-phonenumber-field==7.1.0
django-templated-mail==1.1.1
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
docutils==0.20.1
drf-flex-fields==1.0.2
drf-spectacular==0.26.4
idna==3.4
imagesize==1.4.1
importlib-metadata==7.0.1
importlib-resources==6.0.1
inflection==0.5.1
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.2
phonenumbers==8.13.18
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
snowballstemmer==2.2.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
Sphinx==7.1.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
zipp==3.16.2
sphinxcontrib-openapi==0.8.3
Faker==22.5.0
factory-boy==3.3.0
nltk
python-telegram-bot
django-cors-headers
django-ratelimit
djangorestframework-simplejwt
requests
spacy